requirements.txt
pip>=9
setuptools>=26
wheel>=0.29
pandas
pytest
coverage
flake8
black
yapf
python-gitlab
twine
six==1.15.0
numpy==1.19.5
pandas
scikit-learn==1.0.2
joblib
memoized-property
mlflow
s3fs
gcsfs
google-cloud-storage
termcolor
